The groupwise protocol appears to have changed slightly. This may depend on the specific client/server combinations, of course. For the PCAP that we have access to, the "POST /login" string appears to not be followed by anything.
Fix the ragel to correctly catch the empty "\r\n"
line at the end of the HTTP header.

Comment out the deprecated NM_A_SZ_USERID tag and replace
with the more generic (and possibly problematic) substring
"NM_A_".  Hopefully the combination of the POST line and
this line will be unique to Groupwise.

This protocol would be a good place to tweak the priority
of the match, so that if this were on port 80, the match
quality here would be lower than normal, thus allowing
HTTP port 80 to override it.